chore: restructure relay files
